Potential fix for code scanning alert no. 3: Implicit narrowing conversion in compound assignment

+++ src/main/java/org/mesutormanli/visualnovel/engine/util/RelativeLayout.java
@@ -507,7 +507,7 @@ public class RelativeLayout implements LayoutManager2, java.io.Serializable {
                 Dimension d = component.getPreferredSize();
                 spaceAvailable -= d.height;
             } else {
-                relativeTotal += constraint.doubleValue();
+                relativeTotal += constraint.floatValue();
             }
         }
